refactor: Ticket 字段从 Int 换成 Long

+++ feature-paper-api/src/main/kotlin/plutoproject/feature/paper/api/exchangeshop/ShopCategory.kt
@@ -49,7 +49,7 @@ interface ShopCategory {
     fun addItem(
         id: String,
         itemStack: ItemStack,
-        ticketConsumption: Int = 1,
+        ticketConsumption: Long = 1,
         price: BigDecimal,
         quantity: Int = 1,
         availableDays: List<DayOfWeek> = emptyList()
